Se suman Cambios maria: pasar long a int, completar DAOs de MySQL, agregar Docker

+++ integrador-1/src/main/java/integrador_1/entity/Cliente.java
@@ -8,7 +8,7 @@ import lombok.*;
 @Setter
 @ToString
 public class Cliente {
-    private Long idCliente;
+    private int idCliente;
     private String nombre;
     private String email;
 }
